chore: fix Vercel deploy (npm + build)
