+++ package.json
@@ -36,7 +36,7 @@
     "@types/node": "^20.11.0",
     "@types/react": "^18.2.0",
     "@types/react-dom": "^18.2.0",
-    "@vitejs/plugin-react": "^4.2.1",
+    "@vitejs/plugin-react": "^5.1.2",
     "@vitest/coverage-v8": "^4.0.15",
     "autoprefixer": "^10.4.17",
     "dotenv": "^16.4.5",
@@ -46,6 +46,6 @@
     "tsx": "^4.7.0",
     "typescript": "^5.3.3",
     "vite": "^7.2.6",
-    "vitest": "^1.2.0"
+    "vitest": "^4.0.15"
   }
 }
